+++ client/app/src/main/res/drawable/item_shape.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <solid android:color="#121212"/>
+    <solid android:color="#000000"/>
     <corners android:radius="10dp"/>
     <size android:width="300dp" android:height="100dp"/>
 </shape>
