+++ client/app/src/main/res/layout/fragment_device_list.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="#242424">
+    android:background="#121212">
 
     <TextView
         android:id="@+id/textView3"
