+++ client/app/src/main/res/layout/loading_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="#242424">
+    android:background="#121212">
 
     <TextView
         android:layout_width="wrap_content"
