--- client/app/src/main/res/layout/touchpad_fragment.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/touchpadFragment"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:hapticFeedbackEnabled="true">
-
-</androidx.constraintlayout.widget.ConstraintLayout>
